Check for cache and restore of branches on #19.
When testing on my fork, I had forgotten my fork did not have any of the
prototype branches and for the use case of using GHA on the upstream,
we only check "origin" (the one remote) for now, but that was not as
intended during test that must happen off fork. This testing reminded me
that caching should use the branches as an output and only restore from
cache if they are defined, especially if the default prototype prefix is
used, but nothing is returned, so don't cache empty.