Set last release tag for revision for #48
We still are encountering build failures because the embedded submodule
of usnistgov/OSCAL has the make target we need to get the release assets
to build redirects for, but the BZ2 tarball and ZIP archives default to
include SNAPSHOT without an explicit version. We need to pass that into
the make target as runtime variable when calling it in the
generate_release_assets_redirect script.

https://github.com/usnistgov/OSCAL-Reference/actions/runs/6632414106

Co-Authored-By: Arminta Jenkins <arminta.jenkins@nist.gov>